+++ scripts/lateralus.mixins.js
@@ -230,5 +230,50 @@ define([
     return this;
   };
 
+  /**
+   * @param {Lateralus.Component.Model} Model A constructor, not an instance.
+   * @param {Object} [attributes]
+   * @param {Object} [options]
+   * @return {Lateralus.Component.Model} An instance of the provided Model
+   * constructor.
+   * @method initModel
+   */
+  mixins.initModel = function (Model, attributes, options) {
+    var thisIsLateralus = isLateralus(this);
+
+    var augmentedOptions = _.extend(options || {}, {
+      lateralus: thisIsLateralus ? this : this.lateralus
+    });
+
+    if (!thisIsLateralus) {
+      augmentedOptions.component = this.component || this;
+    }
+
+    return new Model(attributes, augmentedOptions);
+  };
+
+  /**
+   * @param {Lateralus.Component.Collection} Collection A constructor, not an
+   * instance.
+   * @param {Array.<Lateralus.Model>} [models]
+   * @param {Object} [options]
+   * @return {Lateralus.Component.Collection} Am instance of the provided
+   * Collection constructor.
+   * @method initCollection
+   */
+  mixins.initCollection = function (Collection, models, options) {
+    var thisIsLateralus = isLateralus(this);
+
+    var augmentedOptions = _.extend(options || {}, {
+      lateralus: thisIsLateralus ? this : this.lateralus
+    });
+
+    if (!thisIsLateralus) {
+      augmentedOptions.component = this.component || this;
+    }
+
+    return new Collection(models, augmentedOptions);
+  };
+
   return mixins;
 });